* Relocated InProcessCommandBusModule to Cqrs.Ninject package. * Relocated InProcessEventBusModule  to Cqrs.Ninject package. * Relocated InProcessEventStoreModule  to Cqrs.Ninject package. * Relocated InProcessEventStore  to Cqrs package. * Added a constructor to CqrsModule that uses configuration settings instead of parameters.
* Added some extra nuspec tags.
* Renamed some methods and comment in Ninject modules that appear to have been a copy/paste issue.
