+++ Framework/Akka.Net/Cqrs.Akka/Cqrs.Akka.nuspec
@@ -11,7 +11,7 @@
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
 		<description>Use Akka.NET actors in CQRS.NET</description>
 		<copyright>Copyright 2015</copyright>
-		<tags>CQRS CQRS.NET Akka Actors Actor</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices Akka Actors Actor</tags>
 
 		<releaseNotes>
 			Version 2.1
