+++ Framework/Azure/Cqrs.Azure.BlobStorage/Cqrs.Azure.BlobStorage.nuspec
@@ -11,7 +11,7 @@
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
 		<description>Use Azure Blob and Table Storage as the read store and data store in CQRS.NET. This version is targeted at .net 4.0 using WindowsAzure.Storage version 7.2.1</description>
 		<copyright>Copyright 2016</copyright>
-		<tags>CQRS CQRS.NET Azure Blob Storage BlobStorage Table Storage TableStorage</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices Azure Blob Storage BlobStorage Table Storage TableStorage</tags>
 		<releaseNotes>
 			Version 2.0
