+++ Framework/Azure/Cqrs.Azure.ConfigurationManager/Cqrs.Azure.ConfigurationManager.nuspec
@@ -11,7 +11,7 @@
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
 		<description>Use Azure Configuration Manager as the configuration manager in CQRS.NET</description>
 		<copyright>Copyright 2015</copyright>
-		<tags>CQRS CQRS.NET Azure</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices Azure</tags>
 		<releaseNotes>
 			Version 2.2
