+++ Framework/Azure/Cqrs.Azure.DocumentDb/Cqrs.Azure.DocumentDb.nuspec
@@ -2,7 +2,7 @@
 <package>
 	<metadata>
 		<id>Cqrs.Azure.DocumentDb</id>
-		<title>Microsoft Azure DocumentDB for CQRS.NET</title>
+		<title>Microsoft Azure DocumentDB (CosmosDB) for CQRS.NET</title>
 		<version>$version$</version>
 		<authors>Chinchilla Software</authors>
 		<licenseUrl>http://www.apache.org/licenses/LICENSE-2.0</licenseUrl>
@@ -11,7 +11,7 @@
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
 		<description>Use Azure DocumentDB as an event store, read store and data store in CQRS.NET</description>
 		<copyright>Copyright 2014</copyright>
-		<tags>CQRS CQRS.NET Azure DocumentDb DocDb NoSQL</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices Azure DocumentDb CosmosDb Cosmos DocDb NoSQL</tags>
 		<releaseNotes>
 			Version 2.0
