+++ Framework/Azure/Cqrs.Azure.EventHub/Cqrs.Azure.EventHub.nuspec
@@ -11,7 +11,7 @@
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
 		<description>Use Azure EventHubs as a message, event or command bus in CQRS.NET</description>
 		<copyright>Copyright 2016</copyright>
-		<tags>CQRS CQRS.NET Azure ServiceBus EventHub EventHubs</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices Azure ServiceBus EventHub EventHubs</tags>
 		<releaseNotes>
 			Version 2.0
