+++ Framework/Azure/Cqrs.Azure.ServiceBus/Cqrs.Azure.ServiceBus.nuspec
@@ -11,7 +11,7 @@
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
 		<description>Use Azure ServiceBus as a message, event or command bus in CQRS.NET</description>
 		<copyright>Copyright 2015</copyright>
-		<tags>CQRS CQRS.NET Azure ServiceBus</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices Azure ServiceBus</tags>
 		<releaseNotes>
 			Version 2.0
