+++ Framework/Azure/Cqrs.Azure.Storage/Cqrs.Azure.Storage.nuspec
@@ -11,7 +11,7 @@
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
 		<description>Use Azure Blob and Table Storage as the read store and data store in CQRS.NET. This version is targeted at .net 4.5 using WindowsAzure.Storage version 8.0.0</description>
 		<copyright>Copyright 2016</copyright>
-		<tags>CQRS CQRS.NET Azure Blob Storage BlobStorage Table Storage TableStorage</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices Azure Blob Storage BlobStorage Table Storage TableStorage</tags>
 		<releaseNotes>
 			Version 2.0
