+++ Framework/Azure/Cqrs.Azure.WebJobs/Cqrs.Azure.WebJobs.nuspec
@@ -11,7 +11,7 @@
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
 		<description>Use Azure Webjobs with CQRS.NET</description>
 		<copyright>Copyright 2017</copyright>
-		<tags>CQRS CQRS.NET Azure WebJobs</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices Azure WebJobs</tags>
 		<releaseNotes>
 			Version 2.2
