+++ Framework/Cqrs.EventStore/Cqrs.EventStore.nuspec
@@ -11,7 +11,7 @@
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
 		<description>Use Greg Young's event store as the Event Store in CQRS.NET</description>
 		<copyright>Copyright 2013</copyright>
-		<tags>CQRS CQRS.NET EventStore</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices EventStore</tags>
 		<releaseNotes>
 			Version 1.0
