+++ Framework/Cqrs.MongoDB/Cqrs.MongoDB.nuspec
@@ -11,7 +11,7 @@
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
 		<description>Use MongoDB as the read store and data store in CQRS.NET</description>
 		<copyright>Copyright 2016</copyright>
-		<tags>CQRS CQRS.NET Mongo MongoDB NoSQL Ninject</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices Mongo MongoDB NoSQL Ninject</tags>
 		<releaseNotes>
 			Version 2.2
