+++ Framework/Cqrs.Ninject.ServiceHost/Cqrs.Ninject.ServiceHost.nuspec
@@ -11,7 +11,7 @@
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
 		<description>Use Ninject as your IoC container of choice to use WCF with CQRS.NET as ServiceHosts</description>
 		<copyright>Copyright 2013</copyright>
-		<tags>CQRS CQRS.NET ServiceBus Ninject</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices ServiceBus Ninject</tags>
 		<releaseNotes>
 			Version 2.0
