+++ Framework/Cqrs.Sql/Cqrs.Sql.nuspec
@@ -15,7 +15,7 @@
 			use Microsoft SqlServer as an read store and data store in CQRS.NET allowing you to build custom converters to provide mappings from your entities to you SQL table schema structures.
 		</description>
 		<copyright>Copyright 2015</copyright>
-		<tags>CQRS CQRS.NET Sql</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices SQL</tags>
 		<dependencies>
 			<dependency id="Cqrs" version="[$version$]" />
 		</dependencies>
