+++ Framework/Cqrs.Web.Mvc/Cqrs.Web.Mvc.nuspec
@@ -11,7 +11,7 @@
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
 		<description>Use ASP.NET MVC with CQRS.NET</description>
 		<copyright>Copyright 2016</copyright>
-		<tags>CQRS CQRS.NET AspNet MVC AspNetMVC WebHost</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices AspNet MVC AspNetMVC WebHost</tags>
 		<releaseNotes>
 			Version 2.1
