+++ Framework/Cqrs.WebApi/Cqrs.WebApi.nuspec
@@ -11,7 +11,7 @@
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
 		<description>Use ASP.NET Web API with CQRS.NET</description>
 		<copyright>Copyright 2016</copyright>
-		<tags>CQRS CQRS.NET AspNet WebApi AspNetWebApi WebHost</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices AspNet WebApi AspNetWebApi WebHost</tags>
 		<releaseNotes>
 			Version 2.2
 			* Added CORS support with Microsoft.AspNet.WebApi.Cors.
