+++ Framework/Cqrs/Cqrs.csproj
@@ -169,6 +169,7 @@
     <Compile Include="Entities\Entity.cs" />
     <Compile Include="Entities\Range.cs" />
     <Compile Include="Events\DefaultJsonSerializerSettings.cs" />
+    <Compile Include="Events\InProcessEventStore.cs" />
     <Compile Include="Events\SagaEvent.cs" />
     <Compile Include="Events\ISagaEvent.cs" />
     <Compile Include="Events\NotifyEveryoneExceptCallerEventAttribute.cs" />
