+++ Framework/Cqrs/Cqrs.nuspec
@@ -16,7 +16,7 @@
 			CQRS.NET has been designed with modularity in mind. There are a number of supplementary technology packages you can chose from. Every package and design choice made should be interchangeable with custom code if needed.
 		</summary>
 		<copyright>Copyright 2013</copyright>
-		<tags>CQRS CQRS.NET Event-sourcing</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices Event-Sourcing</tags>
 
 		<releaseNotes>
 			Version 2.2
@@ -24,6 +24,7 @@
 			* Added configuration transformations to required app.config and web.config settings are more obvious and set by default.
 			* Refactored CqrsHttpApplication from the Cqrs.WebApi package here without a requirement on SignalR.
 			* Added several base level Hosts for WebAPI, WCF WebJobs and console to base themselves off.
+			* Relocated InProcessEventStore here.
 
 			Version 2.1
