+++ Framework/Ninject/Azure/Cqrs.Ninject.Azure.BlobStorage/Cqrs.Ninject.Azure.BlobStorage.nuspec
@@ -11,7 +11,7 @@
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
 		<description>Use Ninject as your IoC container of choice with Microsoft Azure Blob Storage for CQRS.NET</description>
 		<copyright>Copyright 2016</copyright>
-		<tags>CQRS CQRS.NET Azure Blob Storage BlobStorage Ninject</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices Azure Blob Storage BlobStorage Ninject</tags>
 		<releaseNotes>
 			Version 2.0
