+++ Framework/Ninject/Azure/Cqrs.Ninject.Azure.DocumentDb/Configuration/AzureDocumentDbEventStoreModule.cs
@@ -28,7 +28,7 @@ namespace Cqrs.Ninject.Azure.DocumentDb.Configuration
 		{
 			RegisterFactories();
 			RegisterServices();
-			RegisterCqrsRequirements();
+			RegisterEventStore();
 			RegisterAzureHelpers();
 		}
 
@@ -65,9 +65,9 @@ namespace Cqrs.Ninject.Azure.DocumentDb.Configuration
 		}
 
 		/// <summary>
-		/// Register the all Cqrs command handlers
+		/// Register the <see cref="IAzureDocumentDbEventStoreConnectionStringFactory"/> and <see cref="IEventStore{TAuthenticationToken}"/>
 		/// </summary>
-		public virtual void RegisterCqrsRequirements()
+		public virtual void RegisterEventStore()
 		{
 			Bind<IAzureDocumentDbEventStoreConnectionStringFactory>()
 				.To<AzureDocumentDbEventStoreConnectionStringFactory>()
