+++ Framework/Ninject/Azure/Cqrs.Ninject.Azure.DocumentDb/Configuration/AzureDocumentDbModule.cs
@@ -51,7 +51,7 @@ namespace Cqrs.Ninject.Azure.DocumentDb.Configuration
 		}
 
 		/// <summary>
-		/// Register the all Cqrs command handlers
+		/// Register any CQRS requirements.
 		/// </summary>
 		public virtual void RegisterCqrsRequirements()
 		{
