+++ Framework/Ninject/Azure/Cqrs.Ninject.Azure.DocumentDb/Cqrs.Ninject.Azure.DocumentDb.nuspec
@@ -11,7 +11,7 @@
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
 		<description>Use Ninject as your IoC container of choice with Microsoft Azure DocumentDB for CQRS.NET</description>
 		<copyright>Copyright 2014</copyright>
-		<tags>CQRS CQRS.NET Azure DocumentDb DocDb NoSQL Ninject</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices Azure DocumentDb CosmosDb Cosmos DocDb NoSQL Ninject</tags>
 		<releaseNotes>
 			Version 2.0
