+++ Framework/Ninject/Azure/Cqrs.Ninject.Azure.EventHub.EventBus/Cqrs.Ninject.Azure.EventHub.EventBus.nuspec
@@ -11,7 +11,7 @@
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
 		<description>Use Ninject as your IoC container of choice when using Microsoft Azure Event hubs for CQRS.NET as your event bus</description>
 		<copyright>Copyright 2015</copyright>
-		<tags>CQRS CQRS.NET Azure EventHubs Ninject</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices Azure EventHubs Ninject</tags>
 		<releaseNotes>
 			Version 2.0
