+++ Framework/Ninject/Azure/Cqrs.Ninject.Azure.ServiceBus.CommandBus/Cqrs.Ninject.Azure.ServiceBus.CommandBus.nuspec
@@ -11,7 +11,7 @@
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
 		<description>Use Ninject as your IoC container of choice when using Microsoft Azure ServiceBus for CQRS.NET as your command bus</description>
 		<copyright>Copyright 2015</copyright>
-		<tags>CQRS CQRS.NET Azure ServiceBus Ninject</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices Azure ServiceBus Ninject</tags>
 		<releaseNotes>
 			Version 2.0
