+++ Framework/Ninject/Azure/Cqrs.Ninject.Azure.Wcf/Cqrs.Ninject.Azure.Wcf.nuspec
@@ -13,7 +13,7 @@
 			Use Ninject as your IoC container of choice with Microsoft WCF for CQRS.NET
 		</description>
 		<copyright>Copyright 2017</copyright>
-		<tags>CQRS CQRS.NET Azure WCF Ninject ServiceBus</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices Azure WCF Ninject ServiceBus</tags>
 		<releaseNotes>
 			Version 2.2
