+++ Framework/Ninject/Azure/Cqrs.Ninject.Azure.WebJobs/Cqrs.Ninject.Azure.WebJobs.nuspec
@@ -13,7 +13,7 @@
 			Use Ninject as your IoC container of choice with Microsoft WebJobs for CQRS.NET
 		</description>
 		<copyright>Copyright 2017</copyright>
-		<tags>CQRS CQRS.NET Azure WebJobs Ninject ServiceBus</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices Azure WebJobs Ninject ServiceBus</tags>
 		<releaseNotes>
 			Version 2.2
