+++ Framework/Ninject/Cqrs.Ninject.Akka/Cqrs.Ninject.Akka.nuspec
@@ -2,7 +2,7 @@
 <package>
 	<metadata>
 		<id>Cqrs.Ninject.Akka</id>
-		<title>Ninject IoC container for Akka.Net for CQRS.NET</title>
+		<title>Akka.Net with CQRS.NET and Ninject</title>
 		<version>$version$</version>
 		<authors>Chinchilla Software</authors>
 		<licenseUrl>http://www.apache.org/licenses/LICENSE-2.0</licenseUrl>
@@ -11,7 +11,7 @@
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
 		<description>Use Ninject as your IoC container of choice with Akka.Net for CQRS.NET</description>
 		<copyright>Copyright 2015</copyright>
-		<tags>CQRS CQRS.NET Akka Actors Actor Ninject</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices Akka Actors Actor Ninject</tags>
 		<releaseNotes>
 			Version 2.1
