+++ Framework/Ninject/Cqrs.Ninject.EventStore/Cqrs.Ninject.EventStore.nuspec
@@ -2,7 +2,7 @@
 <package>
 	<metadata>
 		<id>Cqrs.Ninject.EventStore</id>
-		<title>Ninject IoC container for Greg Young's Event Store for CQRS.NET</title>
+		<title>Greg Young's Event Store with CQRS.NET and Ninject</title>
 		<version>$version$</version>
 		<authors>Chinchilla Software</authors>
 		<licenseUrl>http://www.apache.org/licenses/LICENSE-2.0</licenseUrl>
@@ -11,7 +11,7 @@
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
 		<description>Use Ninject as your IoC container of choice with Greg Young's Event Store for CQRS.NET</description>
 		<copyright>Copyright 2013</copyright>
-		<tags>CQRS CQRS.NET EventStore Ninject</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices EventStore Ninject</tags>
 		<releaseNotes>
 			Version 2.0
