+++ Framework/Ninject/Cqrs.Ninject.InProcess.CommandBus/Configuration/InProcessCommandBusModule.cs
@@ -1,84 +1,24 @@
-﻿using System.Linq;
+﻿#region Copyright
+// // -----------------------------------------------------------------------
+// // <copyright company="Chinchilla Software Limited">
+// // 	Copyright Chinchilla Software Limited. All rights reserved.
+// // </copyright>
+// // -----------------------------------------------------------------------
+#endregion
+
+using System;
 using Cqrs.Bus;
 using Cqrs.Commands;
-using Ninject;
 using Ninject.Modules;
 
 namespace Cqrs.Ninject.InProcess.CommandBus.Configuration
 {
 	/// <summary>
-	/// The <see cref="INinjectModule"/> for use with the Cqrs package.
+	/// A <see cref="INinjectModule"/> that configures the <see cref="InProcessBus{TAuthenticationToken}"/> as a <see cref="ICommandPublisher{TAuthenticationToken}"/> and <see cref="ICommandReceiver"/>.
 	/// </summary>
-	public class InProcessCommandBusModule<TAuthenticationToken> : NinjectModule
+	/// <typeparam name="TAuthenticationToken">The <see cref="Type"/> of the authentication token.</typeparam>
+	[Obsolete("Use Cqrs.Ninject.Configuration.InProcessCommandBusModule<TAuthenticationToken> instead.")]
+	public class InProcessCommandBusModule<TAuthenticationToken> : Ninject.Configuration.InProcessCommandBusModule<TAuthenticationToken>
 	{
-		#region Overrides of NinjectModule
-
-		/// <summary>
-		/// Loads the module into the kernel.
-		/// </summary>
-		public override void Load()
-		{
-			RegisterFactories();
-			RegisterServices();
-			RegisterCqrsRequirements();
-		}
-
-		#endregion
-
-		/// <summary>
-		/// Register the all factories
-		/// </summary>
-		public virtual void RegisterFactories()
-		{
-		}
-
-		/// <summary>
-		/// Register the all services
-		/// </summary>
-		public virtual void RegisterServices()
-		{
-		}
-
-		/// <summary>
-		/// Register the all Cqrs command handlers
-		/// </summary>
-		public virtual void RegisterCqrsRequirements()
-		{
-			bool isInProcessBusBound = Kernel.GetBindings(typeof(InProcessBus<TAuthenticationToken>)).Any();
-			InProcessBus<TAuthenticationToken> inProcessBus;
-			if (!isInProcessBusBound)
-			{
-				inProcessBus = Kernel.Get<InProcessBus<TAuthenticationToken>>();
-				Bind<InProcessBus<TAuthenticationToken>>()
-					.ToConstant(inProcessBus)
-					.InSingletonScope();
-			}
-			else
-				inProcessBus = Kernel.Get<InProcessBus<TAuthenticationToken>>();
-
-			Bind<ICommandSender<TAuthenticationToken>>()
-				.ToConstant(inProcessBus)
-				.InSingletonScope();
-
-			Bind<ICommandPublisher<TAuthenticationToken>>()
-				.ToConstant(inProcessBus)
-				.InSingletonScope();
-
-			Bind<IPublishAndWaitCommandPublisher<TAuthenticationToken>>()
-				.ToConstant(inProcessBus)
-				.InSingletonScope();
-
-			Bind<ICommandReceiver<TAuthenticationToken>>()
-				.ToConstant(inProcessBus)
-				.InSingletonScope();
-
-			bool isHandlerRegistrationBound = Kernel.GetBindings(typeof(ICommandHandlerRegistrar)).Any();
-			if (!isHandlerRegistrationBound)
-			{
-				Bind<ICommandHandlerRegistrar>()
-					.ToConstant(inProcessBus)
-					.InSingletonScope();
-			}
-		}
 	}
 }
