+++ Framework/Ninject/Cqrs.Ninject.InProcess.CommandBus/Cqrs.Ninject.InProcess.CommandBus.csproj
@@ -58,6 +58,10 @@
       <Project>{DAFD2A8A-D6B0-4433-8DDF-B98460C64EC6}</Project>
       <Name>Cqrs</Name>
     </ProjectReference>
+    <ProjectReference Include="..\Cqrs.Ninject\Cqrs.Ninject.csproj">
+      <Project>{359bdd3f-785d-4a4a-861f-4951e6e71321}</Project>
+      <Name>Cqrs.Ninject</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(SolutionDir)\.nuget\NuGet.targets" Condition="Exists('$(SolutionDir)\.nuget\NuGet.targets')" />
