+++ Framework/Ninject/Cqrs.Ninject.InProcess.CommandBus/Cqrs.Ninject.InProcess.CommandBus.nuspec
@@ -9,10 +9,14 @@
 		<projectUrl>http://chinchilla-software-com.github.io/CQRS/</projectUrl>
 		<iconUrl>http://chinchilla-software-com.github.io/CQRS/wiki/ChinChilla-Software-Red.png</iconUrl>
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
-		<description>Use Ninject as your IoC container of choice when using an in process channel as your command bus in CQRS.NET</description>
+		<description>Deprecated. Use the Cqrs.Ninject package instead.</description>
 		<copyright>Copyright 2015</copyright>
-		<tags>CQRS CQRS.NET CommandBus Ninject</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices CommandBus Ninject</tags>
 		<releaseNotes>
+			Version 2.2
+
+			* Deprecated. Use Cqrs.Ninject instead.
+
 			Version 2.1
 
 			* #18. Added IPublishAndWaitCommandPublisher to InProcessBus.
@@ -23,6 +27,7 @@
 		</releaseNotes>
 		<dependencies>
 			<dependency id="Cqrs" version="[$version$]" />
+			<dependency id="Cqrs.Ninject" version="[$version$]" />
 		</dependencies>
 	</metadata>
 	<files>
