+++ Framework/Ninject/Cqrs.Ninject.InProcess.EventBus/Cqrs.Ninject.InProcess.EventBus.nuspec
@@ -9,16 +9,21 @@
 		<projectUrl>http://chinchilla-software-com.github.io/CQRS/</projectUrl>
 		<iconUrl>http://chinchilla-software-com.github.io/CQRS/wiki/ChinChilla-Software-Red.png</iconUrl>
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
-		<description>Use Ninject as your IoC container of choice when using an in process channel as your event bus in CQRS.NET</description>
+		<description>Deprecated. Use the Cqrs.Ninject package instead.</description>
 		<copyright>Copyright 2013</copyright>
-		<tags>CQRS CQRS.NET EventBus Ninject</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices EventBus Ninject</tags>
 		<releaseNotes>
+			Version 2.2
+
+			* Deprecated. Use Cqrs.Ninject instead.
+
 			Version 1.0
 
 			* New
 		</releaseNotes>
 		<dependencies>
 			<dependency id="Cqrs" version="[$version$]" />
+			<dependency id="Cqrs.Ninject" version="[$version$]" />
 		</dependencies>
 	</metadata>
 	<files>
