+++ Framework/Ninject/Cqrs.Ninject.InProcess.EventStore/Cqrs.Ninject.InProcess.EventStore.csproj
@@ -54,6 +54,10 @@
       <Project>{DAFD2A8A-D6B0-4433-8DDF-B98460C64EC6}</Project>
       <Name>Cqrs</Name>
     </ProjectReference>
+    <ProjectReference Include="..\Cqrs.Ninject\Cqrs.Ninject.csproj">
+      <Project>{359bdd3f-785d-4a4a-861f-4951e6e71321}</Project>
+      <Name>Cqrs.Ninject</Name>
+    </ProjectReference>
   </ItemGroup>
   <ItemGroup>
     <None Include="Cqrs.Ninject.InProcess.EventStore.nuspec">
