+++ Framework/Ninject/Cqrs.Ninject.InProcess.EventStore/Cqrs.Ninject.InProcess.EventStore.nuspec
@@ -9,16 +9,21 @@
 		<projectUrl>http://chinchilla-software-com.github.io/CQRS/</projectUrl>
 		<iconUrl>http://chinchilla-software-com.github.io/CQRS/wiki/ChinChilla-Software-Red.png</iconUrl>
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
-		<description>Use Ninject as your IoC container of choice when using the built-in, in-memory event store with CQRS.NET</description>
+		<description>Deprecated. Use the Cqrs.Ninject package instead.</description>
 		<copyright>Copyright 2013</copyright>
-		<tags>CQRS CQRS.NET EventStore Ninject</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices EventStore Ninject</tags>
 		<releaseNotes>
+			Version 2.2
+
+			* Deprecated. Use Cqrs.Ninject instead.
+
 			Version 1.0
 
 			* New
 		</releaseNotes>
 		<dependencies>
 			<dependency id="Cqrs" version="[$version$]" />
+			<dependency id="Cqrs.Ninject" version="[$version$]" />
 		</dependencies>
 	</metadata>
 	<files>
