+++ Framework/Ninject/Cqrs.Ninject.Mongo/Configuration/MongoModule.cs
@@ -45,7 +45,7 @@ namespace Cqrs.Ninject.Mongo.Configuration
 		}
 
 		/// <summary>
-		/// Register the all Cqrs command handlers
+		/// Register any CQRS requirements
 		/// </summary>
 		public virtual void RegisterCqrsRequirements()
 		{
