+++ Framework/Ninject/Cqrs.Ninject.Mongo/Cqrs.Ninject.Mongo.nuspec
@@ -2,7 +2,7 @@
 <package>
 	<metadata>
 		<id>Cqrs.Ninject.Mongo</id>
-		<title>Ninject IoC container for legacy MongoDB 2.x for CQRS.NET on .net 4.0</title>
+		<title>Legacy MongoDB 2.x with CQRS.NET and Ninject</title>
 		<version>$version$</version>
 		<authors>Chinchilla Software</authors>
 		<licenseUrl>http://www.apache.org/licenses/LICENSE-2.0</licenseUrl>
@@ -11,7 +11,7 @@
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
 		<description>Use Ninject as your IoC container of choice with MongoDB for CQRS.NET</description>
 		<copyright>Copyright 2013</copyright>
-		<tags>CQRS CQRS.NET Mongo MongoDBs NoSQL Ninject</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices Mongo MongoDBs NoSQL Ninject</tags>
 		<releaseNotes>
 			Version 2.0
