+++ Framework/Ninject/Cqrs.Ninject.MongoDB/Cqrs.Ninject.MongoDB.nuspec
@@ -2,7 +2,7 @@
 <package>
 	<metadata>
 		<id>Cqrs.Ninject.MongoDB</id>
-		<title>Ninject IoC container for MongoDB for CQRS.NET</title>
+		<title>MongoDB with CQRS.NET and Ninject</title>
 		<version>$version$</version>
 		<authors>Chinchilla Software</authors>
 		<licenseUrl>http://www.apache.org/licenses/LICENSE-2.0</licenseUrl>
@@ -11,7 +11,7 @@
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
 		<description>Use Ninject as your IoC container of choice with MongoDB for CQRS.NET</description>
 		<copyright>Copyright 2016</copyright>
-		<tags>CQRS CQRS.NET Mongo MongoDB NoSQL Ninject</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices Mongo MongoDB NoSQL Ninject</tags>
 		<releaseNotes>
 			Version 2.0
