+++ Framework/Ninject/Cqrs.Ninject.WebApi/Cqrs.Ninject.WebApi.nuspec
@@ -11,7 +11,7 @@
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
 		<description>Use Ninject as your IoC container of choice with WebAPI and CQRS.NET</description>
 		<copyright>Copyright 2017</copyright>
-		<tags>CQRS CQRS.NET Ninject WebAPI</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices Ninject WebAPI</tags>
 		<releaseNotes>
 			Version 2.1
