+++ Framework/Ninject/Cqrs.Ninject/Cqrs.Ninject.csproj
@@ -64,6 +64,9 @@
     <Reference Include="Microsoft.CSharp" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Configuration\InProcessCommandBusModule.cs" />
+    <Compile Include="Configuration\InProcessEventBusModule.cs" />
+    <Compile Include="Configuration\InProcessEventStoreModule.cs" />
     <Compile Include="Configuration\ResolvableModule.cs" />
     <Compile Include="Configuration\CqrsModule.cs" />
     <Compile Include="Configuration\MemoryCacheEventStoreModule.cs" />
