+++ Framework/Ninject/Cqrs.Ninject/Cqrs.Ninject.nuspec
@@ -10,12 +10,21 @@
 		<iconUrl>http://chinchilla-software-com.github.io/CQRS/wiki/ChinChilla-Software-Red.png</iconUrl>
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
 		<description>Use Ninject as your IoC container of choice with CQRS.NET</description>
+		<summary>
+			Provides preconfigured Ninject modules that will wire up the required component necessary.
+			
+			Several modules are optional such as the InProcess or the SimplaifySQL modules. You only need to use those if you are using InProcess or SQL.
+		</summary>
 		<copyright>Copyright 2013</copyright>
-		<tags>CQRS CQRS.NET Ninject</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices Ninject</tags>
 		<releaseNotes>
 			Version 2.2
 
 			* Added Host StartUp support.
+			* Relocated InProcessCommandBusModule here.
+			* Relocated InProcessEventBusModule here.
+			* Relocated InProcessEventStoreModule here.
+			* Added a constructor to CqrsModule that uses configuration settings instead of parameters.
 
 			Version 2.1
