+++ Modelling/SolutionTemplate/source.extension.vsixmanifest
@@ -54,7 +54,7 @@ Model commands, events and aggregates visually and let all the handlers and boil
 		<MoreInfo>https://getcqrs.net/</MoreInfo>
 		<Icon>site-icon.png</Icon>
 		<PreviewImage>Tutorial-3-Step-4.png</PreviewImage>
-		<Tags>CQRS CQRS.NET UML Modelling</Tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices UML Modelling</Tags>
 	</Metadata>
 	<Installation>
 		<InstallationTarget Id="Microsoft.VisualStudio.Pro" Version="[11.0,14.0]" />
