+++ Modelling/UmlProfiles/Cqrs.Modelling.nuspec
@@ -16,7 +16,7 @@
 			Model commands, events and aggregates visually and let all the handlers and boiler plate code be generated for you utilising the CQRS.net platform.
 		</summary>
 		<copyright>Copyright 2013</copyright>
-		<tags>CQRS CQRS.NET UML Modelling</tags>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices UML Modelling</tags>
 
 		<releaseNotes>
 			Version 1.601
