+++ packages/repositories.config
@@ -31,6 +31,7 @@
   <repository path="..\Framework\Ninject\Azure\Cqrs.Ninject.Azure.EventHub.EventBus\packages.config" />
   <repository path="..\Framework\Ninject\Azure\Cqrs.Ninject.Azure.ServiceBus.CommandBus\packages.config" />
   <repository path="..\Framework\Ninject\Azure\Cqrs.Ninject.Azure.ServiceBus.EventBus\packages.config" />
+  <repository path="..\Framework\Ninject\Azure\Cqrs.Ninject.Azure.Storage\packages.config" />
   <repository path="..\Framework\Ninject\Azure\Cqrs.Ninject.Azure.Wcf\packages.config" />
   <repository path="..\Framework\Ninject\Azure\Cqrs.Ninject.Azure.WebJobs\packages.config" />
   <repository path="..\Framework\Ninject\Cqrs.Ninject.Akka\packages.config" />
